# docs/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Dashboard",
    "version": "1.0.0",
    "description": "API de dashboard do sistema."
  },
  "servers": [
    {
      "url": "http://localhost:8089",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Payroll",
      "description": "Dados da folha de pagamento"
    }
  ],
  "paths": {
    "/api/v1/dashboard/payroll/{competenceInitial}/{competenceEnd}/summary": {
      "get": {
        "tags": ["Payroll"],
        "summary": "Resumo da folha de pagamento",
        "operationId": "getPayrollSummary",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompetenceDateInitial" },
          { "$ref": "#/components/parameters/CompetenceDateEnd" }
        ],
        "responses": {
          "200": {
            "description": "Resumo da folha de pagamento",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PayrollSummaryResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/dashboard/expense/{competenceInitial}/{competenceEnd}/summary": {
      "get": {
        "tags": ["Expense"],
        "summary": "Resumo das despesas",
        "operationId": "getExpenseSummary",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompetenceDateInitial" },
          { "$ref": "#/components/parameters/CompetenceDateEnd" }
        ],
        "responses": {
          "200": {
            "description": "Resumo das despesas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseSummaryResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "https",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via Keycloak."
      }
    },
    "parameters": {
      "CompetenceDateInitial": {
        "name": "competenceInitial",
        "in": "path",
        "required": true,
        "description": "Competência no formato yyyy-MM-dd",
        "schema": {
          "type": "string",
          "format": "date",
          "example": "2026-06-01"
        }
      },
      "CompetenceDateEnd": {
        "name": "competenceEnd",
        "in": "path",
        "required": true,
        "description": "Competência no formato yyyy-MM-dd",
        "schema": {
          "type": "string",
          "format": "date",
          "example": "2026-06-01"
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Falha na autenticação. Token ausente, inválido ou expirado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 401,
              "error": "Não autorizado",
              "message": "Falha na autenticação. Token ausente, inválido ou expirado",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/dashboard"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Acesso negado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 403,
              "error": "Acesso negado",
              "message": "Você não tem permissão para acessar este recurso",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/dashboard"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "timestamp": "2026-05-03T19:00:00.000Z",
              "status": 500,
              "error": "Erro interno do servidor",
              "message": "Ocorreu um erro inesperado. Por favor, tente novamente mais tarde",
              "path": "/api/v1/dashboard"
            }
          }
        }
      },
      "NotFound": {
        "description": "Não Encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 404,
              "error": "Não Encontrado",
              "message": "Objeto não encontrado",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/dashboard"
            }
          }
        }
      }
    },
    "schemas": {
      "PayrollSummaryResponse": {
        "type": "object",
        "properties": {
          "competenceInitial": { "type": "string", "format": "date", "description": "Data inicial", "example": "2027-02-10" },
          "competenceEnd": { "type": "string", "format": "date", "description": "Data final", "example": "2027-02-10" },
          "entry": {
            "type": "array",
            "description": "Lista as entradas",
            "items": {
              "$ref": "#/components/schemas/EntryTypeResponse"
            }
          }
        }
      },
      "EntryTypeResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID da entrada", "example": 1 },
          "name": { "type": "string", "description": "Nome da entrada", "example": "Provento" },
          "total": { "type": "number", "format": "decimal", "description": "Total da entrada", "example": 4035.00 },
          "events": {
            "type": "array",
            "description": "Lista os eventos",
            "items": {
              "$ref": "#/components/schemas/PayrollEventResponse"
            }
          }
        }
      },
      "PayrollEventResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do evento", "example": 500 },
          "name": { "type": "string", "description": "Nome do evento", "example": "INSS" },
          "total": { "type": "number", "format": "decimal", "description": "Total do evento", "example": 436.00 }
        }
      },
      "ExpenseSummaryResponse": {
        "type": "object",
        "properties": {
          "competenceInitial": { "type": "string", "format": "date", "description": "Data inicial", "example": "2027-02-10" },
          "competenceEnd": { "type": "string", "format": "date", "description": "Data final", "example": "2027-02-10" },
          "categories": {
            "type": "array",
            "description": "Lista de categorias",
            "items": {
              "$ref": "#/components/schemas/ExpenseCategoryResponse"
            }
          }
        }
      },
      "ExpenseCategoryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID da categoria", "example": 1 },
          "name": { "type": "string", "description": "Nome da categoria", "example": "Custo Fixo" },
          "total": { "type": "number", "format": "decimal", "description": "Total da categoria", "example": 2415.00 },
          "events": {
            "type": "array",
            "description": "Lista os despesas",
            "items": {
              "$ref": "#/components/schemas/ExpenseResponse"
            }
          }
        }
      },
      "ExpenseResponse": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Nome da despesa", "example": "Conta de Luz" },
          "total": { "type": "number", "format": "decimal", "description": "Total da despesa", "example": 653.00 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Data e hora do erro em UTC",
            "example": "2026-05-03T19:00:00.000Z"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Descrição resumida do erro HTTP",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Mensagem detalhada do erro",
            "example": "Mensagem de erro"
          },
          "path": {
            "type": "string",
            "description": "Caminho do endpoint que gerou o erro",
            "example": "/api/v1/dashboard"
          }
        }
      }
    }
  }
}